+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.1] - 22/02/2025
+
+### Fixed
+- Fixed navigation when there are page and size parameters in the URL
+- Fixed max letters ins filter by material
+- Fixed indentation and improve layout in `filterByColor`
+- Fixed maxLength for string trimming and update layout for filter components
+
 ## [1.8.0] - 22/02/2025
 
 ### Fixed
@@ -279,6 +287,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new version of the project. With products and services.
 
+[1.8.1]: https://github.com/CristhianDaza/megaprom/pull/30/files
 [1.8.0]: https://github.com/CristhianDaza/megaprom/pull/29/files
 [1.7.6]: https://github.com/CristhianDaza/megaprom/pull/28/files
 [1.7.5]: https://github.com/CristhianDaza/megaprom/pull/27/files